Refining class refactor
some relabeling. Moved where buffer was added to. Removed length checks
for prompt detection and used buffer instead of data. changed callback
function names and added the end even message to sshObj. Changed main
public functions so it could be run as an instance of the class or as
the static method on the class.